Rename DeliveryChannelPolicies folders to DeliveryChannels
