+++ src/protagonist/API/Features/DeliveryChannels/Requests/CreateDeliveryChannelPolicy.cs
@@ -1,4 +1,4 @@
-﻿using API.Features.DeliveryChannelPolicies.Validation;
+﻿using API.Features.DeliveryChannels.Validation;
 using API.Infrastructure.Requests;
 using DLCS.Core;
 using DLCS.Model.Policies;
@@ -6,7 +6,7 @@ using DLCS.Repository;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
-namespace API.Features.DeliveryChannelPolicies.Requests;
+namespace API.Features.DeliveryChannels.Requests;
 
 public class CreateDeliveryChannelPolicy : IRequest<ModifyEntityResult<DeliveryChannelPolicy>>
 {
