+++ src/protagonist/API/Features/DeliveryChannels/Requests/GetDeliveryChannelPolicy.cs
@@ -4,7 +4,7 @@ using DLCS.Repository;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
-namespace API.Features.DeliveryChannelPolicies.Requests;
+namespace API.Features.DeliveryChannels.Requests;
 
 public class GetDeliveryChannelPolicy: IRequest<FetchEntityResult<DeliveryChannelPolicy>>
 {
